fix: remove dead test helpers and guard canvas metadata early
